cEP-0001: Remove BLD
The concept of the BLD was based on suggestions from several community members requesting centralized leadership at that time.

I believe coala has grown up and that the concept of the BLD is no longer needed.

There is a diverse set of very capable people, maintainers are starting to meet regularly to give coala a clear direction. Our democratic processes are maturing and our community is able to make better decisions than a single person could do.